WIP: add trace diff view HTTP endpoint and shared SVG package
- New endpoint GET /api/v2/traces/diff/view serves HTML page with
  embedded SVG tree, pan/zoom, tooltips, and inline form
- Landing page at /api/v2/traces/diff/view (no params) with form UI
- Extract SVG rendering into pkg/tracediffsvg for reuse by CLI and server
- Refactor CLI trace-diff-svg to use shared package
- Update docs with browser view and SVG CLI sections
